requirements.txt
speedtest_cli==2.1.2
pandas==1.0.3
matplotlib==3.1.2
numpy==1.17.4
